+++ src/session-client/context.js
@@ -15,9 +15,15 @@ class Context {
    * of script resources which should be managed by OM SDK service.
    * @param {!Partner} partner
    * @param {?Array<!VerificationScriptResource>} verificationScriptResources
+   * @param {?string=} contentUrl The provided content url needed if OMSDK
+   * is running in cross-domain. Defaults to null.
+   * @param {?string=} customReferenceData Arbitrary reference data the
+   * integrator can share with verification scripts. Has no effect in
+   * mobile app environment. Defaults to null.
    * @throws error if the supplied partner is undefined or null.
    */
-  constructor(partner, verificationScriptResources) {
+  constructor(partner, verificationScriptResources, contentUrl = null,
+    customReferenceData = null) {
     argsChecker.assertNotNullObject('Context.partner', partner);
 
     this.partner = partner;
@@ -26,7 +32,12 @@ class Context {
     this.slotElement = null;
     /** @type {?HTMLVideoElement} */
     this.videoElement = null;
+    /** @type {?string} */
+    this.contentUrl = contentUrl;
+    /** @type {?string} */
+    this.customReferenceData = customReferenceData;
   }
+
   /**
    * Specifies the video element within the WebView.
    * @param {?HTMLVideoElement} videoElement The video element