+++ test/unit/session-client/media-events-test.js
@@ -1,14 +1,14 @@
-goog.module('omid.test.sessionClient.VideoEvents');
+goog.module('omid.test.sessionClient.MediaEvents');
 
 const AdSession = goog.require('omid.sessionClient.AdSession');
 const argsChecker = goog.require('omid.common.argsChecker');
 const VastProperties = goog.require('omid.common.VastProperties');
-const VideoEvents = goog.require('omid.sessionClient.VideoEvents');
+const MediaEvents = goog.require('omid.sessionClient.MediaEvents');
 const {InteractionType, VideoPlayerState, VideoPosition} = goog.require('omid.common.constants');
 const {asSpy} = goog.require('omid.test.typingUtils');
 
-/** @type {!VideoEvents} */
-let videoEvents;
+/** @type {!MediaEvents} */
+let mediaEvents;
 
 /** @type {!AdSession} */
 let mockAdSession;
@@ -21,14 +21,14 @@ let mockAdSession;
  */
 function testForwardingFunction(methodName, f, ...args) {
   it('should be relayed to the service', () => {
-    f.apply(videoEvents, args);
+    f.apply(mediaEvents, args);
 
     expect(mockAdSession.sendOneWayMessage)
         .toHaveBeenCalledWith(methodName, ...args);
   });
 }
 
-describe('VideoEventsTest', () => {
+describe('MediaEventsTest', () => {
   beforeEach(() => {
     mockAdSession = jasmine.createSpyObj(
         'AdSession', [
@@ -41,7 +41,7 @@ describe('VideoEventsTest', () => {
     spyOn(argsChecker, 'assertNumber');
     spyOn(argsChecker, 'assertNumberBetween');
 
-    videoEvents = new VideoEvents(mockAdSession);
+    mediaEvents = new MediaEvents(mockAdSession);
   });
 
   describe('constructor', () => {
@@ -49,12 +49,12 @@ describe('VideoEventsTest', () => {
       expect(argsChecker.assertNotNullObject).toHaveBeenCalledWith(
           'MediaEvents.adSession', mockAdSession);
     });
-    it('should throw an error if ad session already has video events', () => {
+    it('should throw an error if ad session already has media events', () => {
       asSpy(mockAdSession.registerMediaEvents).and.callFake(() => {
         throw new Error();
       });
 
-      expect(() => new VideoEvents(mockAdSession)).toThrow();
+      expect(() => new MediaEvents(mockAdSession)).toThrow();
     });
   });
 
@@ -63,17 +63,17 @@ describe('VideoEventsTest', () => {
       const vastProperties = new VastProperties(
           false /* isSkippable */, 0 /* skipOffset */,
           false /* isAutoPlay */, VideoPosition.PREROLL /* position */);
-      videoEvents.loaded(vastProperties);
+      mediaEvents.loaded(vastProperties);
 
       expect(argsChecker.assertNotNullObject).toHaveBeenCalledWith(
-        'MediaEvents.loaded.vastProperties', vastProperties);
+          'MediaEvents.loaded.vastProperties', vastProperties);
     });
 
     it('should be relayed to the service', () => {
       const vastProperties = new VastProperties(
           false /* isSkippable */, 0 /* skipOffset */,
           false /* isAutoPlay */, VideoPosition.PREROLL /* position */);
-      videoEvents.loaded(vastProperties);
+      mediaEvents.loaded(vastProperties);
 
       expect(mockAdSession.sendOneWayMessage).toHaveBeenCalledWith(
           'loaded', vastProperties);
@@ -82,44 +82,44 @@ describe('VideoEventsTest', () => {
 
   describe('start', () => {
     it('should verify duration by calling the argsChecker', () => {
-      const videoEvents = new VideoEvents(mockAdSession);
-      videoEvents.start(5, .5);
+      const mediaEvents = new MediaEvents(mockAdSession);
+      mediaEvents.start(5, .5);
 
       expect(argsChecker.assertNumber).toHaveBeenCalledWith(
-        'MediaEvents.start.duration', 5);
+          'MediaEvents.start.duration', 5);
     });
 
     it('should verify mediaPlayerVolume by calling the argsChecker', () => {
-      const videoEvents = new VideoEvents(mockAdSession);
-      videoEvents.start(5, .5);
+      const mediaEvents = new MediaEvents(mockAdSession);
+      mediaEvents.start(5, .5);
 
       expect(argsChecker.assertNumberBetween).toHaveBeenCalledWith(
           'MediaEvents.start.mediaPlayerVolume', .5, 0, 1);
     });
 
     testForwardingFunction(
-        'start', VideoEvents.prototype.start, 5 /* duration */,
+        'start', MediaEvents.prototype.start, 5 /* duration */,
         0.6 /* volume */);
   });
 
   describe('volumeChange', () => {
     it('should verify mediaPlayerVolume by calling the argsChecker', () => {
-      const videoEvents = new VideoEvents(mockAdSession);
-      videoEvents.volumeChange(.6);
+      const mediaEvents = new MediaEvents(mockAdSession);
+      mediaEvents.volumeChange(.6);
 
       expect(argsChecker.assertNumberBetween).toHaveBeenCalledWith(
           'MediaEvents.volumeChange.mediaPlayerVolume', .6, 0, 1);
     });
 
     testForwardingFunction(
-        'volumeChange', VideoEvents.prototype.volumeChange, 0.6 /* volume */);
+        'volumeChange', MediaEvents.prototype.volumeChange, 0.6 /* volume */);
   });
 
   describe('playerStateChange', () => {
     it('should verify playerState by calling the argsChecker', () => {
-      const videoEvents = new VideoEvents(mockAdSession);
+      const mediaEvents = new MediaEvents(mockAdSession);
       const videoPlayerState = VideoPlayerState.FULLSCREEN;
-      videoEvents.playerStateChange(videoPlayerState);
+      mediaEvents.playerStateChange(videoPlayerState);
 
       expect(argsChecker.assertNotNullObject).toHaveBeenCalledWith(
           'MediaEvents.playerStateChange.playerState', videoPlayerState);
@@ -127,14 +127,14 @@ describe('VideoEventsTest', () => {
 
     testForwardingFunction(
         'playerStateChange',
-        VideoEvents.prototype.playerStateChange,
+        MediaEvents.prototype.playerStateChange,
         VideoPlayerState.FULLSCREEN);
   });
 
   describe('adUserInteraction', () => {
     it('should verify adUserInteraction by calling the argsChecker', () => {
-      const videoEvents = new VideoEvents(mockAdSession);
-      videoEvents.adUserInteraction(InteractionType.CLICK);
+      const mediaEvents = new MediaEvents(mockAdSession);
+      mediaEvents.adUserInteraction(InteractionType.CLICK);
 
       expect(argsChecker.assertNotNullObject).toHaveBeenCalledWith(
           'MediaEvents.adUserInteraction.interactionType',
@@ -143,45 +143,45 @@ describe('VideoEventsTest', () => {
 
     testForwardingFunction(
         'adUserInteraction',
-        VideoEvents.prototype.adUserInteraction,
+        MediaEvents.prototype.adUserInteraction,
         InteractionType.CLICK);
   });
 
   describe('firstQuartile', () => {
     testForwardingFunction(
-        'firstQuartile', VideoEvents.prototype.firstQuartile);
+        'firstQuartile', MediaEvents.prototype.firstQuartile);
   });
 
   describe('midpoint', () => {
-    testForwardingFunction('midpoint', VideoEvents.prototype.midpoint);
+    testForwardingFunction('midpoint', MediaEvents.prototype.midpoint);
   });
 
   describe('thirdQuartile', () => {
     testForwardingFunction('thirdQuartile',
-        VideoEvents.prototype.thirdQuartile);
+        MediaEvents.prototype.thirdQuartile);
   });
 
   describe('complete', () => {
-    testForwardingFunction('complete', VideoEvents.prototype.complete);
+    testForwardingFunction('complete', MediaEvents.prototype.complete);
   });
 
   describe('pause', () => {
-    testForwardingFunction('pause', VideoEvents.prototype.pause);
+    testForwardingFunction('pause', MediaEvents.prototype.pause);
   });
 
   describe('resume', () => {
-    testForwardingFunction('resume', VideoEvents.prototype.resume);
+    testForwardingFunction('resume', MediaEvents.prototype.resume);
   });
 
   describe('bufferStart', () => {
-    testForwardingFunction('bufferStart', VideoEvents.prototype.bufferStart);
+    testForwardingFunction('bufferStart', MediaEvents.prototype.bufferStart);
   });
 
   describe('bufferFinish', () => {
-    testForwardingFunction('bufferFinish', VideoEvents.prototype.bufferFinish);
+    testForwardingFunction('bufferFinish', MediaEvents.prototype.bufferFinish);
   });
 
   describe('skipped', () => {
-    testForwardingFunction('skipped', VideoEvents.prototype.skipped);
+    testForwardingFunction('skipped', MediaEvents.prototype.skipped);
   });
 });